third_party/WebKit/LayoutTests/fast/dom/dir-auto-insert-text-invalidation-expected.txt

Style invalidation when inserting RTL text descendant of dir=auto

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(outerAuto).direction is "ltr"
PASS getComputedStyle(textInsertionPoint).direction is "ltr"
PASS getComputedStyle(affectedByAuto).direction is "ltr"
PASS getComputedStyle(explicitLTR).direction is "ltr"
PASS internals.updateStyleAndReturnAffectedElementCount() is 4
PASS getComputedStyle(outerAuto).direction is "rtl"
PASS getComputedStyle(textInsertionPoint).direction is "rtl"
PASS getComputedStyle(affectedByAuto).direction is "rtl"
PASS getComputedStyle(explicitLTR).direction is "ltr"

Forcing everything back to LTR with dir attribute:

PASS internals.updateStyleAndReturnAffectedElementCount() is 4
PASS getComputedStyle(outerAuto).direction is "ltr"
PASS getComputedStyle(textInsertionPoint).direction is "ltr"
PASS getComputedStyle(affectedByAuto).direction is "ltr"
PASS getComputedStyle(explicitLTR).direction is "ltr"
PASS successfullyParsed is true

TEST COMPLETE

؄ل
